requirements.txt
streamlit==1.25.0
matplotlib==3.8.0
langchain==0.3.29
langchain-community==0.3.14
boto3==1.28.0
openai==0.29.0
pandas==2.1.0
numpy==1.25.1
python-dotenv==1.0.0
faiss-cpu==1.7.4
requests==2.31.0
kaggle==1.5.13
